chore(CreateNewPost): nest client component in a server component to prepare for route guard
